Increase detection timestamp precision
Previously, detection timestamp was taken in the main thread, which
could be delayed due to things like display redrawing etc.

Now, we take crossing timestamp directly in the "IRQ" handler, which
wiringPi runs in a high-priority thread. Moreover, we also take the
timestamp of the next round to be the same as object detection
timestamp. Previously, new timestamp was taken for that.